Fix MOSFET symbol pinout
This fixes the mistake in the MOSFET symbol that incorrectly paired
up input pins 1-2 with output pins 5-6 and input 3-4 with 7-8.
Instead, pins 1-2 (source & gate) map to output pins 7-8 (drain) and
3-4 map to 5-6.

The result of this mistake in v01 of the board means that every pair
of display rows are flipped. At a high refresh rate this should not
be visible, but it does require a workaround in the driver software
that swaps the contents of every pair of display rows.